Database System
Decided to remove the CacheManager system for this project. Adds unnecessary headaches for the project.
Changed how the database stores dates to make it easier to read.
Animals now load when the program loads.
Changed the variable types in the database and Animal.java to be strings in all cases except the animalID. Makes it easier to manipulate.
Updated all other files to reflect the new database setup.