Update Database.java
Removed the executeQuery and buildQuery methods, as they were only used one time, so I put the functionality into the one method that needed it.
Removed loadAnimals and loadKennels in favor of loading everything in one call, using one connection. The loadAll method would create a connection, load the animals, close that connection, open a new connection, load the kennels, and close that connection. Now everything is loaded with one connection, back to back.
Simplified as much code as I could in Database.java.