Employee System
Created Employee class, only gets used when a person logs in.
Added 'Name' and 'Manager' fields to Users table.
Disabled the Dashboard ButtonBar if you haven't logged in yet.
If the logged in User is a manager, a "Management" button appears on the dashboard.
Added exception throwing to all database methods that create a connection.
Added a currentUser variable to the Database class, which is set when someone logs in.
Added a test manager account, Username: Manager Password: Manager
Added a test employee account, Username: Employee Password: Employee